Remove extracted util files add them as static methods in respective classes
